Rework DataCoordinate custom containers.
This renames some existing classes, so it is a breaking API change,
but one whose effects should be minimal to nonexistent outside
daf_butler.  Overall it aligns the hierarchy better with the built-in
typing hierarchy, injecting some interfaces that had implementations
but no ABCs, and it cleans up the handling of mutability by separating
frozenset from set views.