Stop trying to cache DimensionRecord lookup failures.
We've historically treated DimensionRecord lookup failures as a
non-error (we just don't return records we didn't find), at least
mostly because we allow implied dimension dependencies to be None
(e.g. an exposure need not have a physical_filter).  But that's
actually broken because of DM-21840, and I now think the right
solution there is to just make those foreign keys NOT NULL after all.

That's a long way of saying that I think we could _probably_ start
treating DimensionRecord lookup failures as errors, and simplify a lot
of code in the process (though a lot of test would break).  I'm not
confident enough to try that now and I don't want to deal with fixing
the tests, but I am confident enough to say that lookup failures are
so rare that we *can* simplify by not caching them, and that's what
this commit does.

This is relevant because I'm about to reimplement the caching entirely
in terms of the new DimensionRecord containers, and I don't want those
to have to worry about caching failed lookups.  But I wanted to make
the (slight) behavioral change a separate commit (this one) from the
refactoring one.