Rewrite and reorganize data ID expansion/standardization.
Before this change, we had two big blocks of not-really-cooperative
code for standardizing dicts into DataCoordinates:

 - DataCoordinate.standardize did as much as it could without access
   to a Registry to fetch DimensionRecords from;

 - Registry.expandDataId could do everything, but it's slow
   (many-small-queries latency issues) and tied to Registry, and it
   couldn't really rely on DataCoordinate.standardize as much as it'd
   like to, because DataCoordinate.standardize throws away information
   it can't use even if Registry.expandDataId might have been able to
   use it.

With the new DimensionRecord containers, we can now give
DataCoordinate.standardize the ability to attach records and expand
implied dimensions via records itself, by fetching those records from
a container it's passed.  That's delegated to a new low-level method
on HeterogeneousDimensionRecordAbstractSet that has most of the logic
from Registry.expandDataId.  Registry.expandDataId itself now just
creates a caching DimensionRecord container and calls
DataCoordinate.standardize.

I've also unified the implementations of DataCoordinate.standardize
and DataCoordinate.subset by moving them into a new
DataCoordinateCommonState class.  This separates the logic for
determining what state (dimensions identified, expansion flags) the
new data ID can have from the act of actually constructing them, which
will be useful in later commits we want to apply these operations to
many DataCoordinate objects with the same state.

This adds a new method to Registry and SqlRegistry without a
RemoteRegistry implementation, and that's why it isn't decorated with
abstractmethod yet.  It'll be easier to fix that after we make the new
DimensionRecord containers serializable.

Finally, this adds a check to DataCoordinate.standardize that any
passed keys are actually dimension names; this addresses a
long-standing problem where users would get a keyword argument to some
higher-level API wrong, and it would be forwarded as **kwargs down to
DataCoordinate.standardize and then silently ignored.  And it turns
out we were doing that even in our own test utility code!