Add logic to recognize non-standard keys in DataCoordinate.
This reimplements some of the special handling of non-standard keys
from Butler._findDatasetRef in the hopes of being able to move it all
down to Registry (and thus work on many more interfaces).  But it's
just a start at that; I realized while trying to make
Butler._findDatasetRef use the new code that we really need to make
queryDatasets work on CALIBRATION collections first.  But I think what
I've done so far will still be useful eventually, so I'm keeping it.