Add custom containers for DatasetRefs.
I'm not completely happy with this - lots of classes, lots of
boilerplate, and even a bit of logic duplication - but everything I
try to do to fix it seems to make it worse in some other respect, and
I've spent far too much time tinkering with it already.  But I'm
pretty happy with the ABCs, and perhaps more importantly I think
they're pretty close to the minimal versions of the interfaces we
need, which makes them something good to build on.