Add "split" transfer mode
This mode allows an ingest to have some files that are inside
the datastore already and some that are outside. This is a common
scenario for execution butler where raws are in a shared location
outside of a datastore but calibrations and other products are
inside the butler datastore.

"auto" mode will choose the "split" option in that case.

This mode is effectively transfer=None for files inside the datastore
and transfer=direct for those outside the datastore.