Remove softlink option in cache retrieval
Hard links protect against races of other processes doing cache
expiry just after the soft link has been created. A hard link
solves this because either the hard link is created or it isn't
created.

If it is created the file in the exempt directory can be used
without risk of being deleted prematurely and it is safe to
delete the version in the main cache. If the hard link can't
be created because another process has just deleted the
file, it will be as if the file wasn't in the cache at all.

Assumes that hard links will work on any file system we use
for the cache.