Join in a dummy subquery when dataset queries are known to fail.
Bailing out early caused some downstream code that assembles the rest
of the code to fail.  This is still a bit messy; the right fix is to
refactor so the objects that represent no-result queries don't even
have SQLAlchemy column objects embedded in them.  But that's beyond the
scope of this ticket.  It's one of the things I'll try to do on
DM-31725.