Rework handling of governor dimension restrictions.
It took me a while to figure out why the set-like
GovernorDimensionRestriction class kept needing new set-op method
variants for ~every context in which it was being used, but the answer
is that it was really a mapping of sets (one for each governor
dimension), and hence when instances have different dimensions
(mapping keys) dimensions, their relationship is not set-like.

The best solution seems to be to remove that class in favor of the
kind of mapping-of-sets that backed it, and move its convenience
functionality that's important for CollectionSummary to
CollectionSummary itself.

I changed some attribute names and the module name for collecton
summary to get them consistent with now-preferred style; it made sense
to do this at the same time because I wanted to find all code using
CollectionSummary in order to make sure it was updated appropriately,
and renames made MyPy very good at that.  I also switched from
Dimension instance keys to `str` dimension name keys, since I think
we'll be doing more of that with RFC-834 adopted.

Finally, I've changed the logic in the user-expression checking code
to track all dimensions, and then filter down to just the governor
dimensions at the boundary with the rest of the query system.  This
just seemed the best for overall code simplicity.