Keep track of artifact existence in transfer_from
We check for dataset existence once in Butler.transfer_from
to filter out missing datasets that we don't want to go in
the registry. Then later on we check inside datastore
to see which files exist because composite disassembly
means that we can't know up front whether a file
should be there or not.