Enable datastore.trash to trash files it doesn't know about
datastore.trash() is optimized to scan the records table and
quickly move those records to the trash table for later
deletion.

For execution mode the files might be there but the records
won't be.  Therefore, to allow consistency for dataset pruning,
when trust mode is enabled the file system is examined and files
removed if they are present even if there are no records.