Separate dataset dimensions from requested dimensions in queries.
The dimensions passed directly to queryDataIds were always intended to
be the dimensions the caller wanted back, which may not be the same as
the dimensions used to constrain the query.  We'd partially handled
that distinction in the past, but we had been rolling dataset
constraint dimensions into the requested set.  This was most obvious
in queryDimensionRecords, which didn't quite handle this properly, and
hence couldn't handle constraints on datasets that have dimensions
beyond the dimension records' own.  Fixing queryDataIds to track
dataset-constraint dimensions separately solves that problem, too.