Avoid string types with ClassVar
As of Python 3.9.8, this does not work with Pydantic BaseModel, and
does not work in a way that causes the whole module to fail to
import.  See https://github.com/samuelcolvin/pydantic/issues/3401.

In general, Pydantic does not work well with string types as enabled
by from __future__ import annotations, which is why making that
feature the default has been delayed.  See
https://lwn.net/Articles/858576/ for more information.

Fix this by removing from __future__ import annotations from the
files using ClassVar and explicitly quote the type arguments that
contain forward references (all of which are outside the scope of
what Pydantic cares about).