Empty kwargs during dataId rewriting
Previously non-dimensions were removed from kwargs and
valid dimensions were retained. None of this agreed
with the definition of "unused" in the docstring
since those now raise an exception (and were
already removed from kwargs). It's much more
consistent to transfer kwargs over to the new
dataId and empty kwargs.