Check all dimensions, not just missing ones
When trying to match up record values to a specific
dimension, there are sometimes cases where the dimension
is given as well as record values. In that case we need
to also assign the records to a dimension for later
resolution. Without doing this the records look
like unrecognized keywords.