Ignore keys return values in named mappings
types.Mapping expects .keys() to return KeysView but
these classes return an AbstractSet. The set is compatible
with KeysView to the extent we need but KeysView is
a concrete class and not an abstract one so it is not
desirable to involve it. For now this means we ignore
the warning.