Fix compatibility with moto 3.0
moto3 broke the s3 tests solely in test_butler.py. It seems
that something has changed in using the mock_s3 class decorator
such that the inherited methods (the ones that are doing all
the testing) are not actually using the mocks at all but are
using raw boto. Moving the start and stop calls explicitly
into setUp and tearDown fixes the problem.