First attempt to improve exceptions raised by registry.
This introduces exception class hierarchy for use in registry methods. It
is limited to registry only and does not touch core module or datastore,
though some changes in Butler class were needed to make it consistent
with registry exceptions. Registry docstrings were updated to enumerate
expected exception types. There is still no guarantee that raised
exceptions will be limited to classes in this hierarchy, there is always
a possibility that exceptions from other modules will leak. In
particular there was no attempt in this commit to convert possible
SQLAlchemy exceptions into registry-specific types.