Overhaul dataset type wildcards and queries.
This adds a new DatasetTypeWildcard class for initial processing of
wildcard expressions and moves the logic for actually resolving those
wildcards into the manager classes.  Dataset type and collection
handling now have roughly analogous code structure.

This also changes the return type of queryDatasetTypes from Iterator
to Iterable, as per RFC-878, and emits deprecation warnings when
patterns or unregistered dataset types are passed to queryDataIds or
queryDimensionRecords, as per RFC-879.