Handle no-results queries better in findDatasets follow-up.
This effectively requires a very minor API change: the return type
for DataCoordinateQueryResults is now DatasetQueryResults, the base
class of the previous return type, because the previous return type
can't handle the case where the DatasetType doesn't exist.  I'd be
somewhat surprised if there was any code outside of daf_butler tests
and QuantumGraph generation calling findDatasets at all, and I'd be
shocked if there was any dependence on the current return type, so I
think this is okay.