Refactor dataset subquery logic.
This used to be one massive method; now it's two medium-large methods
and two very simple ones, and it's much easier to follow.

This shouldn't change the query behavior at all, but it does collapse
some UNION ALL constructs by merging their IN clauses when they are
otherwise the same - the original code had an unnecessary distinction
between CALIBRATION and other collection types (as well as some
necessary distinctions, which this preserves).