Introduce new exception for missing dataset types.
This inherits from KeyError as well as DatasetTypeError to preserve
the behavior of existing catch blocks, and it seems somewhat logical
anyway (even if LookupError might be slightly better than KeyError).
It's theoretically possible for this change to break existing catch
blocks that trapped for DatasetTypeError first and then KeyError, but
I'm pretty sure that's just hypothetical.