Rework sorting of collections in query results and scripts.
This removes sorting from Registry.queryCollections itself, moving it
in a limited way to various CLI scripts, adjusting tests accordingly.

The behavior of queryCollections w.r.t. order is now documented: the
order is unspecified, except for the order of children of the same
parent collection, when flattenChains=True.  This retroactively makes
usage in ctrl_mpexec safe (which was originally lucky, and recently
became incorrect after more aggressive sorting was added).  The
structure of the implementation makes it very hard to maintain
search order for children of chains while sorting everything else, and
maintaining search order for children of chains is more important.

Script methods now sort collection results to make the output
deterministic and easier to read.  Children of chained collections are
already split off and handled separately in most query-collections
modes, so search order is now preserved there (but chains=flatten has
unspecified order).