Join in a dummy dataset subquery even when the query is doomed.
This prevents query construction from failing when someone queries for
a skypix-dimension dataset type in collections where none exist.  When
the datasets do exist, we rely on the dataset table to bring in the
skypix dimension, since those don't have tables of their own, and we
need to do the same when the query is doomed to avoid triggering an
exception about skypix dimensions not being directly joinable.