Add serialization support for datastore records (DM-33493)
Adds a bunch of pydantic classes for datastore records and implements
serialization of the records in `Quantum` class. I switched back
`QuantumProvenanceData` to be a pydantic model to use the same pydantic
classes for datastore records. Moved `DatastoreRecordData` class (and its
pydantic friend) to a separate module, for some reason pydantic was not
happy when I added `SerializedDatastoreRecordData` class to the
`datastore` module.