Allow timespan selects/constraints from non-CALIBRATION collections.
The timespan for a dataset subquery for a non-CALIBRATION collection
is hereby defined to be a fully unbounded timespan, which will let us
UNION those subqueries against CALIBRATION-collection queries for the
same dataset before doing a temporal join against a dimension like
exposure or visit.

It would make sense to change the behavior of queryDatasetAssociations
to associate non-CALIBRATION collections with a full timespan instead
of None at the same time, but that's an API change that I don't want
to tackle right now.  Eventually I'd like to replace
queryDatasetAssocations entirely with a method-chaining interface to
get associations from queryDatasets, and I think that RFC wil be the
time to tackle this.