Allow timespan selects/constraints from non-CALIBRATION collections.
The timespan for a dataset subquery for a non-CALIBRATION collection
is hereby defined to be NULL, which will let us UNION those subqueries
against CALIBRATION-collection queries for the same dataset before
doing a temporal join against a dimension like exposure or visit.

We'll have to do those joins carefully, with an "is NULL" check that
makes a NULL timespan join on the dataset side match against any
temporal dimension record.