Add option to compute collection rank in dataset subqueries.
This provides an alternate implementation for constructing an important
aspect of find-first dataset queries, but it isn't actually used yet.

This case-based approach is much closer to Andy Salnikov's initial
implementation of find-first dataset queries; I think it's likely that
changing it was at most unnecessary performance-wise (I was looking for
ways to simplify a query the planner was having trouble with, but this
construct - despite being hard for humans to parse - probably wasn't
playing a role in that).  And some other query system changes have made
it preferable from a code-organization standpoint, providing its no
worse from a performance standpoint (ideally, there would be no
difference, as the query planner would recognize the approaches as
identical, but I don't want to assume that).