Replace fromSelectable with from_columns in topological DB classes.
SQLAlchemy is moving towards being much more strict about the
difference between the columns of a SELECT statement and the columns
of a FromClause, such as a table or an aliased subquery, and in the
future we won't be able to access a ".columns" across the board (e.g.
on Select objects, it'll be "exported_columns" instead).