Integrate relations with QueryBuilder and dimension managers.
Most of this ticket is a major overhaul of QueryBuilder and the
DimensionRecordStorage classes to use relations throughout, with
conversion back to the old SimpleQuery system now happening only at
the boundary between QueryBuilder and Query.

Ideally, that'd be plenty for one commit, but doing that work revealed
some preexisting problems with our dimension caching system and how we
test for data ID validity in a few different contexts.  But I couldn't
fix those in earlier commits without the relation machinery introduced
on this one, and I couldn't get that new relation machinery to pass
tests without fixing them.  So here they are.

The first problem was that all the previous cleverness about
populating the CachingDimensionRecordStorage cache one row at a time
never made sense: it's much more efficient to just fetch all rows for
cached dimensions whenever any of them are requested, since by
definition we only configure caching for dimensions whose record count
is small enough that we can easily hold them all in memory, and
probably small enough that fetching all of them is still dominated by
per-query overheads.

Changing that fixes a bug in which QueryDimensionRecordStorage (used
exclusively for 'band') happily returned nonexistent dimension records
if you gave it a nonexistent data ID, and hence you could expand a bad
data ID and not notice.  Or, more precisely, this commit sidesteps
that still-existing problem, because 'band' is actually configured to
use CachingDimensionRecordStorage on top of
QueryDimensionRecordStorage, and now CachingDimensionRecordStorage
prevents the bad behavior.  In a few commits I plan to remove the
fetch method from all of those storage objects, and then the bug will
really fully be gone.

Anyhow, fixing that bug creates new problems, in the form of a
Registry test that was actually checking for the wrong behavior - we
*can* tell when a query with a bad band value is doomed, because we
try to expand the data ID up front and (now) that fails.  But we don't
want an exception for this - we want a doomed query.  So that
snowballed into additional logic changes involving exceptions
vs. doomed queries.

Last but not least, this modifies CachingDimensionRecordStorage and
GovernorDimensionRecordStorage to allow them to have an unset (None)
cache, and removes the refresh methods for populating those caches.
Calling higher-level refresh methods now actually unsets the caches,
forcing them to be repopulated on first use.  It didn't actually make
sense to have both refresh and clearCaches before, but we need
clearCaches in order to have something we can call at transaction
rollback without hitting the database; this is exactly the problem
identified (with collection caching) on DM-35498.  Also, having
clearCaches make the cache an empty container was a bug waiting to
happen for GovernorDimensionRecordStorage, since that (usually
incorrectly) would say that there are no governor dimension values in
the database.