Add tie breaker for ambiguous visit
A visit that is defined using day_obs/seq_num has an ambiguity
in that it is possible for the one-to-one visit and the
seq_start/seq_end visit to have the same values for day_obs/seq_num.

Now explicitly catch an ambiguous visit and use the default
visit system from the instrument record to determine which one
is likely to be the desired one.