Rework timespan SQL expression classes.
This removes the analogous expression class for spatial regions, as
well as the base class they shared.  We need an abstraction layer
above SQLAlchemy's for timespans, since those can have a
representation that involves multiple columns in some databases, and
SQLAlchemy's ColumnElement can't handle that.  But our regions are
always just an opaque blob column, so that abstraction layer wasn't
carrying its weight, despite the nice symmetry between regions and
timespans in other respects.  A new classmethod constructor has been
added to FieldSpec to take care of the only important thing this class
did - holding the defaults for how region columns should be defined.

That results in a number of abstract methods now being first defined
in TimespanDatabaseRepresentation.  For the most part, these are just
transfers from the old base class with slightly adjusted docs, but
there are a few small updates:

- fromLiteral now accepts None as a way to support NULL expressions;

- fromSelectable has been replaced from_columns, since SQLAlchemy is
  moving towards being much more strict about the difference between
  the columns of a SELECT statement and the columns of a FromClause,
  such as a table or an aliased subquery, and in the future we won't
  be able to access a ".columns" across the board (e.g.  on Select
  objects, it'll be "exported_columns" instead).

- we now use tuples instead of generators for a few methods, since
  generators for 1- or 2-element results involves a lot more overhead
  with no real gain.