Automatically generate implied dimension records in test utility code.
This required a fair amount of restructuring of the test utility code
implementation.  Some variable names were changed as well to better
reflect nomenclature used elsewhere.

This required some changes to unit tests because it's not possible to
distinguish between cases where the caller wants some dimension records
created automatically and cases where the caller defined a data ID
that is inconsistent with existing records - the logic now just creates
records as necessary to make the given data ID valid.