improve CLI option & argument caching
Previously we were caching the command line arguments in the
context object for downstream use, but the format of the arguments
at this point is too raw; arguments were split on whitespace
but not separators (like "--foo=bar").

This adds a helper fucntion that uses the click parser to parse
the arguments, and then stores the parsed option names, option
values, and argument values for use by downstream code.