Add a way to mark which dimensions populate others.
In the future, this will be used to make it easier to export dimension
data without knowing the details of the dimension combinations that
represent many-to-main joins (DM-34838).

There's no real gain to updating the dimensions configuration before
that happens, but merging these changes early is useful because it
will make software versions with just this change much better able to
handle future data repositories that use them in their dimensions
configuration, even if the associated functionality isn't available as
a result.