Use Relation and QueryBackend for dataset queries.
This is the first of several incremental changes to the query system,
which won't really pay off until the end (when we'll be able to remove
all the old stuff that Relation replaces).

Note that the new methods on QueryBackend here are for the most part
transfers of code (certainly transfers of logic) from QueryBackend,
though it's all been updated at least slightly to make use of
Relation.