Clean up some str/DimensionElement interfaces.
This allows DimensionRecordStorageManager to accept str keys, which is
important to help us move in the RFC-834 direction of preferring str
keys.

It also changes the implementation of a few methods that already accept
either str or Named instances to use getattr with a default value
instead of isinstance; this should be faster.  This does require us to
cast a bit to satisfy MyPy, but only because MyPy has to care about
hypothetical str subclasses that have a .name attribute, and we just
don't.