Disable move export mode and assume direct == None for export
move should never be allowed for export since the files must
still stay with the original datastore.

direct is also dangerous since we do not want files that
could be moved out of the datastore on import and we do not
want a direct import to have files disappear if the original
butler no longer wants the file.