Add factories for common predicates.
These are not used yet, but they will be soon (and we'll be adding
more, too, as needed).  They involve a bit more boilerplate than
their equivalents in the current query system, in large part because
the relation system is going to want to be able to serialize these, and
that means they need to be more self-describing.