Modernize imports and annotations in dataset storage.
This was one of the few places in daf_butler still using absolute
imports.

It also includes reducing imports from 'typing' in favor of builtins
and collections.abc, including using | instead of Union and Optional,
in these modules only, because I'm about to modify them heavily and
didn't want to write the old style or leave files with both styles in
them.