Add ColumnTypeInfo.
This includes information for all of the columns whose types can vary
between registries as a result of different configuration.  It can
thus only be constructed by something that knows about the Registry
manager classes, but it's useful to have just this much information in
a class available to core, since those types do affect objects there.

The name of this class doesn't make a lot of sense right now, but that
will change in a few commits.