Handle single-time arguments in Timespan.overlaps.
(and TimespanDatabaseRepresentation.overlaps).

This makes the Python interface more consistent with the string
expression language, which uses OVERLAPS for timespan-time arguments,
and it will sidestep a problem with function name resolution in the new
daf_relation expression system, since normally that looks in the
operator module first and a method second, and "contains" (but not
"overlaps") is a member of the operator module.