Integrate relations into dataset subqueries and storage managers.
This is the first of several commits that replace pieces of Registry
query code with new versions that are based on the new daf_relation
package.  The real value of that switch won't be realized until the
query system as a whole is replaced (and in particular the Query class
itself is wholly rewritten), and in the meantime there will be a fair
amount of transitional code tying the new and old interfaces together.

This moves the former DatasetRecordStorage.find implementation into
SqlRegistry, since it can now delegate to
DatasetRecordStorage.make_relation.  And once we have a QueryBackend
implementation for RemoteRegistry, we can move it further into the
base class Registry and use the same implementation for all
registries.