Use relations to rewrite Query and QueryResults.
This is a full rewrite of Query - it's now a concrete final class
instead a hierarchy (extensibility is provided by the QueryContext and
QueryBackend interfaces it delegates to), and it's all backed by
Relation instead of interacting with SQLAlchemy objects at all.

That's important for future client-server butler work, where the
expectation is that Query can be used more-or-less as-is with new
QueryContext and QueryBackend implementation (with the caveat that an
abstract interface rarely survives the definition of its second-ever
derived class unscathed).

It's also where Relation really starts to show its power in making the
query system _conceptually_ simpler, even if there's still a ton of
code: before, when applying a new operation (e.g. sorting) to an
existing query, we basically had to try to reconstruct all of the
parameters that went into the original query, then modify them
according to the new operation (and any that had been applied since).
Effectively every Query factory method had to know about all of the
things that could go into construting a Query.  Now, we can sometimes
just append a new relation operation to the existing tree, and even
the more complex cases can be expressed as relatively straightforward
operations on that existing tree.

That gives us the backbone we'll need for real query composability;
it's not quite here yet only because I'm focusing now on
reimplementing existing interfaces rather than defining newer more
powerful ones.  But it's a pretty small jump from here to a lot of
things we've wanted for a long time, especially QG generation,
including:

- data ID set uploads, implemented as iteration-engine leaf relations that
  get transferred to the SQL engine;

- adding new dimensions to an existing query with customization of the
  spatiotemporal relationships to use, which is just a frontend to
  QueryBackend.make_dimension_relation;

- vectorizing calibration and other prerequisite lookups, which is a
  little more code in `Query.find_datasets` to add temporal joins.
  And the previously-thorny problem how how to represent the results
  of that - which need to involve rows that represent both a
  DatasetRef and dimensions that aren't part of that ref's DatasetType
  - is neatly solved by direct iteration over a Query instead of
  DatasetQueryResults.

On that note, the other big change here is that pretty much all of the
functionality in the QueryResults classes has moved to Query; the
QueryResults classes are just thin facades now, and
DimensionRecordQueryResults now delegates directly to Query (as its
data ID and dataset siblings do) instead of the query-factory stuff it
had before.

The changes in tests here mostly reflect the fact that using relation
improves our ability to do diagnostics on queries that return no
results as well as our ability to see when a query is doomed before
execution.  I'm a bit worried the diagnostics in some cases could also
be getting a bit less readable to end-users with this change, though,
because they're coming from daf_relation now and use its terminology.
But I think they'll still carry the important pieces of information
and they'll be much more precise, so let's see how they work in the
wild before trying to intercept and rewrite them in daf_butler.