Relax exception message checks in registry tests.
We'll soon get different exception messages for missing governor
dimension values depending on whether the offending value appears
first in a data ID that's expanded or a query WHERE string.  That's
reasonable behavior (since the exception type is the same, and each
message is useful in its own right), and that upcoming change makes it
easier to ensure handling of other bad dimension values is consistent.
But it make it happen we first need to relax the test on the exception
message to the common denominator.