Permit data IDs passed to query methods to remain unexpanded.
Data ID expansion can fail when a given data ID value is missing from
the database, but when those bad data ID values appear anywhere else
in a query (or a bad data ID is already fully-expanded), the query
system's behavior is to raise only for governor dimensions and to
return no results otherwise.  In order to make that behavior
consistent, we need to allow data ID expansion to fail and proceed with
the unexpanded data ID and a doomed query.

This commit only allows unexpanded data IDs in the lower levels of the
query systems; the higher level Registry methods still call
expandDataId without guarding against its failure.  So this is just
setting the stage for more DM-31725 (or spinoff thereof) changes in
the future.