Add support for exposure regions.
Exposure records do not include region, to get region we match exposure
with a visit (any visit) and use region from that visit. It only works
if a visit is already defined when per-exposure datasets are inserted.
This does not work for raw datasets, which are inserted before the visits
are defined based on the same raws.