Use datastore transactions in butler methods that trash datasets
Butler does not know that a registry transaction is sufficient.

This also add a transaction parameter to bridge moveToTrash method
to make use of that new transaction context.

Without this there is no way for the ephemeral bridge to roll
back a transaction from a failure in registry. This works for
the normal bridge because that is also using a shared
database transaction.