Simplify materialized skypix overlaps.
We were maintaining a lot of unused, probably too-clever code for
dealing with spatial overlaps between regular database dimension
elements and arbitrary skypix dimensions, while only overlaps with
the common skypix dimension are actually supported everywhere else.

This commit drops that extra complexity, and with it an undesirable
dependency on GovernorDimensionRecordStorage.values (which I'd like to
remove) and some unnecessary table locking.