Remove test of registry disassembly
This test was there for when registry would disassemble
and give datastore individual components. We no longer
do this so there is no need to test the functionality,
especially since that relied on special logic where
the read and write storage classes did not match.
Now we have storage class conversion infrastructure
that test is no longer valid and datastore can just
depend on whether a component has been requested
without trying to guess from the storage class
discrepancy that this has been an external
disassembly.