Move component handling out of dataset manager.
findDataset and getDatasetType are now the only registry methods that
support components; removing support from those would cause a lot of
inconvenience downstream, and the complexity cost in Registry for
keeping it there is quite low. But we want to get it out of the
manager classes to get rid of that unpleasant copy.copy on the storage
object, and to keep component support from accidentally spreading to
other places where we don't want to maintain it.

At the same time, I've changed the exception raised for missing
dataset types in the manager to be MissingDatasetTypeError for
consistency with the query methods, but adding KeyError as a base
class of that exception to maintain backwards compatibility.