Remove now-redundant component export test.
Originally this test appeared to test that component datasets could be
exported directly; that's what the docstring still said prior to this
commit, though it was actually just exporting the parent datasets and
querying for the component datasets on import.  Now querying for the
components is deprecated, so this test is redundant with a few others
that already check basic export-import round-trip.